Modified Analytics Page
- separated the components (charts, cards)
- fixed the scrolling problem (scrollbar is now in the main div)
- fixed the overflowing prob of user categories circles
- sticky title header
- fixed the gap size
